requirements.txt
<<<<<<< HEAD
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# Data fetching and processing
yfinance>=0.2.0
ta>=0.10.0  # Technical analysis library

# Machine Learning
torch>=1.9.0
tensorflow>=2.10.0
scikit-learn>=1.0.0
joblib>=1.2.0

# Sentiment Analysis
tweepy>=4.0.0  # Twitter API
praw>=7.0.0    # Reddit API
textblob>=0.17.1
nltk>=3.8.1

# Visualization and Analysis
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pytz>=2021.1
schedule>=1.2.0

# Testing
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
anyio>=4.0.0

# Monitoring and Logging
prometheus_client>=0.17.0  # For metrics
python-telegram-bot>=20.0  # For alerts

# Backtesting
backtrader>=1.9.76.123

# Model Optimization and Tracking
optuna>=3.0.0  # For hyperparameter optimization
mlflow>=2.8.0  # For experiment tracking
shap>=0.42.0   # For model explainability

# Web and data processing
beautifulsoup4>=4.12.0
pyarrow>=12.0.0  # For parquet files
fastparquet>=2023.4.0  # Alternative parquet engine

# For TFT model
tensorflow-probability>=0.19.0
=======
pandas
pyyaml
yfinance
pytest
pytest-mock
textblob
nltk
python-dotenv
requests
tweepy
praw
>>>>>>> e719f64 (news_sentiment, support for REDDIT+TWITTER news + tests)
